backend/requirements.txt
#########################
# 基础 Web 服务
#########################
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart==0.0.20
httpx[http2]>=0.27.0

#########################
# 数据验证 / 配置管理
#########################
pydantic>=2.5.0
python-dotenv>=1.0.0

#########################
# 数据库
#########################
SQLAlchemy==2.0.40
PyMySQL==1.1.1
asyncmy>=0.2.9    # 异步 MySQL 驱动

#########################
# Redis
#########################
redis>=5.0.1
aioredis>=2.0.1

#########################
# 阿里云短信服务
#########################
aliyun-python-sdk-dysmsapi==1.0.0
